feat: add public nuguard.redteam.report module (to_markdown / to_json) (#44)
Extract the private _findings_to_markdown, _scenario_coverage_table,
_append_remediation_plan, and _truncate_evidence helpers from
nuguard/cli/commands/redteam.py into a new public module at
nuguard/redteam/report.py, mirroring the existing nuguard.behavior.report
interface.

Public API:
- to_markdown(findings, meta, remediation_plan, scenario_records) -> str
- to_json(findings, meta, remediation_plan) -> str

The CLI functions are replaced with thin delegators so existing behaviour
is unchanged.  NuGuard-app and other callers can now import the report
utilities directly without going through CLI internals.
